+++ backend/app/schemas/review.py
@@ -5,7 +5,7 @@ from datetime import datetime
 
 class ReviewBase(BaseModel):
     restaurant_id: int
-    rating: float = Field(..., ge=1, le=5, description="Рейтинг от 1 до 5")
+    rating: int = Field(..., ge=1, le=5, description="Рейтинг от 1 до 5")
     comment: Optional[str] = None
 
 
@@ -14,7 +14,7 @@ class ReviewCreate(ReviewBase):
 
 
 class ReviewUpdate(BaseModel):
-    rating: Optional[float] = Field(None, ge=1, le=5, description="Рейтинг от 1 до 5")
+    rating: Optional[int] = Field(None, ge=1, le=5, description="Рейтинг от 1 до 5")
     comment: Optional[str] = None
